% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Acos}
\alias{Acos}
\title{Element-wise inverse cosine on a Elemental matrix}
\usage{
Acos(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Replaces each entry from MatrixA by its inverse cosine function
}

