% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Asin}
\alias{Asin}
\title{Element-wise inverse sine on a Elemental matrix}
\usage{
Asin(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Replaces each entry from MatrixA by its inverse sine function
}

