% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Atan}
\alias{Atan}
\title{Element-wise inverse tangent on a Elemental matrix}
\usage{
Atan(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Replaces each entry from MatrixA by its inverse tangent function
}

