% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Axpy}
\alias{Axpy}
\title{Multiplies Matrix X by scalar alpha and adds Matrix Y}
\usage{
Axpy(alpha, MatrixX, MatrixY)
}
\arguments{
\item{alpha}{Scalar that multiplies each element of MatrixX}

\item{MatrixX}{An Elemental matrix (input)}

\item{MatrixY}{An Elemental matrix (input and output)}
}
\value{
None
}
\description{
Performs the operation `Y = aX + Y` on Elemental matrices
}

