% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ColSwap}
\alias{ColSwap}
\title{Swaps two columns in an Elemental matrix}
\usage{
ColSwap(MatrixA, to, from)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}

\item{to}{col destination (input) 0-indexed}

\item{to}{col origin (input) 0-indexed}
}
\value{
None
}
\description{
Interchanges two columns inside an Elemental matrix MatrixA
}

