% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Conjugate}
\alias{Conjugate}
\title{Calculates the conjugate transpose of an Elemental matrix}
\usage{
Conjugate(MatrixA)
}
\arguments{
\item{MatrixA}{An elemental Matrix (input, output)}
}
\value{
None
}
\description{
It replaces the matrix by its complex conjugate.
}

