% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Copy}
\alias{Copy}
\title{Copies MatrixA into MatrixB}
\usage{
Copy(MatrixA, MatrixB)
}
\arguments{
\item{MatrixA}{An elemental Matrix (input)}

\item{MatrixB}{An elemental Matrix (output)}
}
\value{
None
}
\description{
Performs the operation `B = A` using Elemental Matrices
}

