% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Cos}
\alias{Cos}
\title{Element-wise cosine on a Elemental matrix}
\usage{
Cos(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Replaces each entry from MatrixA by its cosine function
}

