% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Demmel}
\alias{Demmel}
\title{Generates a Demmel Elemental Matrix}
\usage{
Demmel(MatrixA, n)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}

\item{n}{the dimension of the matrix (input)}
}
\value{
None
}
\description{
Generates an n x n demel Elemental matrix.
}

