% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{DiagonalScale}
\alias{DiagonalScale}
\title{Multiplies a diagonal matrix by an Elemental matrix}
\usage{
DiagonalScale(side, Matrixd, MatrixX, orientation = "NORMAL")
}
\arguments{
\item{side}{"L" or "R" (input), indicating whether the diagonal matrix is on the 
left or right side}

\item{Matrixd}{an Elemental 1-D matrix (input)}

\item{orientation}{indicates whether MatrixX is transposed or no (input).
Possible values "N" or "T" Normal or Transposed}

\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Multiplies the diagonal matrix made from the vector Matrixd by MatrixX, 
storing the result on MatrixX. Which is equivalent to:
}
\details{
MatrixX = diag(Matrixd) %*% MatrixX
}

