% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{DiagonalSolve}
\alias{DiagonalSolve}
\title{Solves a diagonal linear system}
\usage{
DiagonalSolve(side, uplo, Matrixd, MatrixX, offset, orientation = "NORMAL")
}
\arguments{
\item{side}{"L" or "R" (input), indicating whether the diagonal matrix is on the 
left or right side}

\item{uplo}{(input). It can take the values "U" or "L" to indicate what part of
matrixX has entries.}

\item{Matrixd}{an Elemental 1-D matrix (input)}

\item{MatrixX}{an Elemental matrix (input, output)}

\item{offset}{an integer value that indicates the diagonal offset of MatrixX}

\item{orientation}{indicates whether MatrixX is transposed or no (input).
Possible values "N" or "T" Normal or Transposed}
}
\value{
None
}
\description{
Solves a linear system with a diagonal matrix diag( Matrixd ) and right hand
sides MatrixX. It stores the result on MatrixX.
}

