% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ElDistMatrix-class}
\alias{ElDistMatrix}
\alias{ElDistMatrix-class}
\title{Class to represent an Elemental Distributed Matrix}
\description{
Class to represent an Elemental Distributed Matrix
}
\section{Slots}{

\describe{
\item{\code{ptr}}{A pointer to the distributed matrix object (C++)}

\item{\code{datatype}}{A character that stores the datatype}
}}

