% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ElDistPermutation-class}
\alias{ElDistPermutation}
\alias{ElDistPermutation-class}
\title{Class to represent an Elemental Distributed Permutation (for pivoting)}
\description{
Class to represent an Elemental Distributed Permutation (for pivoting)
}
\section{Slots}{

\describe{
\item{\code{ptr}}{A pointer to the permutation object (C++)}
}}

