% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ElGrid-class}
\alias{ElGrid}
\alias{ElGrid-class}
\title{Class to represent the Grid}
\description{
Class to represent the Grid
}
\section{Slots}{

\describe{
\item{\code{ptr}}{A pointer to the grid (C++)}
}}

