% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ElMatrix-class}
\alias{ElMatrix}
\alias{ElMatrix-class}
\title{Class to represent an Elemental Matrix}
\description{
Class to represent an Elemental Matrix
}
\section{Slots}{

\describe{
\item{\code{ptr}}{A pointer to the matrix object (C++)}

\item{\code{datatype}}{A character that stores the datatype}
}}

