% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ElPermutation-class}
\alias{ElPermutation}
\alias{ElPermutation-class}
\title{Class to represent an Elemental Sequential Permutation (for pivoting)}
\description{
Class to represent an Elemental Sequential Permutation (for pivoting)
}
\section{Slots}{

\describe{
\item{\code{ptr}}{A pointer to the permutation object (C++)}
}}

