% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Exp}
\alias{Exp}
\title{Element-wise exponential on a Elemental matrix}
\usage{
Exp(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Replaces each entry from MatrixA by its exponential function
}

