% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Fill}
\alias{Fill}
\title{Fills an Elemental Matrix with a scalar value}
\usage{
Fill(MatrixA, alpha)
}
\arguments{
\item{MatrixA}{(input, output)}

\item{alpha}{a scalar value (input)}
}
\value{
None
}
\description{
Fills each element from the MatrixA with the given scalar value
}

