% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{FillDiagonal}
\alias{FillDiagonal}
\title{Fills the diagonal elements in an Elemental Matrix with a scalar value}
\usage{
FillDiagonal(MatrixA, alpha, offset = 0)
}
\arguments{
\item{MatrixA}{(input, output)}

\item{alpha}{a scalar value (input)}

\item{offset}{an integer value that indicates the diagonal offset}
}
\value{
none
}
\description{
Fills each element in the diagonal from MatrixA with the given scalar value
}

