% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Environment.R
\name{Finalize}
\alias{Finalize}
\title{Finalize}
\usage{
Finalize()
}
\description{
Finalizes MPI
}
\details{
This function must be called at the end of the script (at least at the end
of the distributed part).
}

