% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Gemm}
\alias{Gemm}
\title{Matrix matrix multiplication of Elemental matrices}
\usage{
Gemm(orientationA = "NORMAL", orientationB = "NORMAL", alpha, MatrixA,
  MatrixB, beta, MatrixC)
}
\arguments{
\item{orientationA}{indicates whether matrixA is transposed or no (input).
Possible values "N" or "T" Normal or Transposed}

\item{orientationB}{indicates whether matrixB is transposed or no (input).
Possible values "N" or "T" Normal or Transposed}

\item{alpha}{Scalar that multiplies the matrix vector product(input)}

\item{MatrixA}{an Elemental matrix (input)}

\item{MatrixB}{an Elemental matrix (input)}

\item{beta}{Scalar that multiplies scales Matrixy (input)}

\item{MatrixC}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Performs the operation `C = aAB + bC` on Elemental matrices
}

