% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Ger}
\alias{Ger}
\title{Outer product of of Elemental matrices (vectors)}
\usage{
Ger(alpha, Matrixx, Matrixy, MatrixA)
}
\arguments{
\item{alpha}{Scalar that multiplies the outer product (input)}

\item{Matrixx}{a 1-D Elemental matrix (input)}

\item{Matrixy}{a 1-D Elemental matrix (input)}

\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Performs the operation `A = alpha xy^H + A` on Elemental matrices
}

