% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Geru}
\alias{Geru}
\title{Outer product of of Elemental matrices (vectors)}
\usage{
Geru(alpha, Matrixx, Matrixy, MatrixA)
}
\arguments{
\item{alpha}{Scalar that multiplies the outer product (input)}

\item{Matrixx}{a 1-D Elemental matrix (input)}

\item{Matrixy}{a 1-D Elemental matrix (input)}

\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Performs the operation `A = alpha xy^T + A` on Elemental matrices
}

