% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Hadamard}
\alias{Hadamard}
\title{Calculates the element-wise multiplication of two matrices}
\usage{
Hadamard(MatrixA, MatrixB, MatrixC)
}
\arguments{
\item{MatrixA}{An Elemental matrix (input)}

\item{MatrixB}{An Elemental matrix (input)}

\item{MatrixC}{An Elemental matrix (input and output)}
}
\value{
None
}
\description{
Performs the element-wise operation `C = A * B` on Elemental matrices
}

