% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Hemm}
\alias{Hemm}
\title{Hermitian matrix matrix multiplication of Elemental matrices}
\usage{
Hemm(side, uplo, alpha, MatrixA, MatrixB, beta, MatrixC)
}
\arguments{
\item{side}{"L" or "R" (input), indicating whether matrixA is on the 
left or right side}

\item{uplo}{(input). It can take the values "U" or "L" to indicate what part of
matrixA has entries.}

\item{alpha}{Scalar that multiplies the matrix vector product(input)}

\item{MatrixA}{an Elemental matrix (input)}

\item{MatrixB}{an Elemental matrix (input)}

\item{beta}{Scalar that multiplies scales Matrixy (input)}

\item{MatrixC}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Performs the operation `C = aAB + bC` or `C = aBA + bC` on Elemental matrices,
wehere A is implicitly Hermitian
}

