% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Her2k}
\alias{Her2k}
\title{Hermitian rank 2k update of Elemental matrices}
\usage{
Her2k(uplo, orientation, alpha, MatrixA, MatrixB, beta, MatrixC)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" to indicate what part of
matrixA has entries.}

\item{orientation}{indicates whether matrixA is an adjoint or no (input).
Possible values "N", "A" Normal or Adjoint}

\item{alpha}{Scalar that multiplies the matrix products (input)}

\item{MatrixA}{an Elemental matrix (input)}

\item{MatrixB}{an Elemental matrix (input)}

\item{beta}{Scalar that multiplies scales Matrixy (input)}

\item{MatrixC}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Performs the operation `C = a AB^H + a_conj BA^H + bC` or 
`C = a A^H B + a_conj B^H A + bC` on Elemental matrices, only the
 selected triangle of C is modified
}

