% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Herk}
\alias{Herk}
\title{Hermitian rank k update of Elemental matrices}
\usage{
Herk(uplo, orientation, alpha, MatrixA, beta, MatrixC)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" to indicate what part of
matrixA has entries.}

\item{orientation}{indicates whether matrixA is an adjoint or no (input).
Possible values "N", "A" Normal or Adjoint}

\item{alpha}{Scalar that multiplies the matrix product(input)}

\item{MatrixA}{an Elemental matrix (input)}

\item{beta}{Scalar that multiplies scales Matrixy (input)}

\item{MatrixC}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Performs the operation `C = aAA^H + bC` or `C = aA^H A + bC` on Elemental matrices,
only the selected triangle of C is modified
}

