% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{HilbertSchmidt}
\alias{HilbertSchmidt}
\title{Performs the Hilbert Schmidt inner product}
\usage{
HilbertSchmidt(MatrixA, MatrixB)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input)}

\item{MatrixB}{an Elemental matrix(input)}
}
\value{
the value of the inner product
}
\description{
Performs the Hilbert Schmidt inner product
}

