% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Identity}
\alias{Identity}
\title{Initializes an Elemental matrix as an identity matrix}
\usage{
Identity(MatrixA, rows, cols)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input and output)}

\item{rows}{the number of rows (input)}

\item{cols}{the number of columns (input)}
}
\value{
None
}
\description{
Initializes an Elemental matrix as an identity matrix
}

