% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ImagPart}
\alias{ImagPart}
\title{Extract the imaginary part of a complex Elemental matrix}
\usage{
ImagPart(MatrixA, MatrixAIm)
}
\arguments{
\item{MatrixA}{An Elemental matrix (input)}

\item{MatrixAIm}{An Elemental matrix (input, output)}
}
\value{
None
}
\description{
Takes the imaginary part from MatrixA and stores it intro MatrixAIm
}

