% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Environment.R
\name{Initialize}
\alias{Initialize}
\title{Initialize}
\usage{
Initialize()
}
\description{
Initializes MPI allowing to use Distributed Matrices
}
\examples{
Initialize()
A<-DistMatrix()
Uniform(A,2,2)
Print(A)
Finalize()
}

