% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MakeReal}
\alias{MakeReal}
\title{Makes an Elemental (complex)matrix real}
\usage{
MakeReal(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Makes the matrix A real, keeping only its real part.
}

