% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MakeTrapezoidal}
\alias{MakeTrapezoidal}
\title{Makes an Elemental matrix trapezoidal}
\usage{
MakeTrapezoidal(uplo, MatrixA, offset = 1)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" to indicate what part of
the matrix should be kept.}

\item{MatrixA}{an Elemental matrix (input, output)}

\item{offset}{an integer that indicates the diagonal offset}
}
\value{
None
}
\description{
Makes the matrix A trapezoidal, keeping only its lower (or upper) 
triangular part with a given offset.
}

