% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Max}
\alias{Max}
\title{Finds the maximum value in an Elemental matrix}
\usage{
Max(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input)}
}
\value{
$value the maximum value in the matrix

  	   $indices an array with the coordinates of the maximum value
}
\description{
Finds both, the maximum scalar value in the matrix and its coordinates (row, col)
}

