% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MaxAbs}
\alias{MaxAbs}
\title{Finds the maximum absolute value in an Elemental matrix}
\usage{
MaxAbs(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input)}
}
\value{
$value the maximum absolute value in the matrix

  	   $indices an array with the coordinates of the maximum value
}
\description{
Finds both, the maximum absolute scalar value in a 
matrix and its coordinates (row, col)
}

