% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Min}
\alias{Min}
\title{Finds the minimum value in an Elemental matrix}
\usage{
Min(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input)}
}
\value{
$value the minimum value in the matrix

  	   $indices an array with the coordinates of the minimum value
}
\description{
Finds both, the minimum scalar value in the matrix and its coordinates (row, col)
}

