% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{MinAbs}
\alias{MinAbs}
\title{Finds the minimum absolute value in an Elemental matrix}
\usage{
MinAbs(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input)}
}
\value{
$value the minimum absolute value in the matrix

  	   $indices an array with the coordinates of the minimum value
}
\description{
Finds both, the minimum absolute scalar value in a 
matrix and its coordinates (row, col)
}

