% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{QuasiTrsv}
\alias{QuasiTrsv}
\title{Quasi-Triangular solve for Elemental matrices}
\usage{
QuasiTrsv(uplo, orientation, MatrixA, Matrixx)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" to indicate what part of
matrixA has entries.}

\item{orientation}{indicates whether matrixA is transposed or no (input).
Possible values "N" or "T" Normal or Transposed}

\item{MatrixA}{an Elemental matrix (input)}

\item{Matrixx}{a 1-D Elemental matrix (input, output)}
}
\value{
None
}
\description{
Performs the operation `x = A^(-1)x` on Elemental matrices.
}

