% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RealPart}
\alias{RealPart}
\title{Extract the real part of a complex Elemental matrix}
\usage{
RealPart(MatrixA, MatrixAReal)
}
\arguments{
\item{MatrixA}{An Elemental matrix (input)}

\item{MatrixAReal}{An Elemental matrix (input, output)}
}
\value{
None
}
\description{
Takes the real part from MatrixA and stores it intro MatrixAReal
}

