% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{RElem}
\alias{RElem}
\alias{RElem-package}
\title{Provide wrappers to Elemental (C++) functions}
\description{
This package allows to connect R with the high performance library
Elemental.
}
\details{
It enables to do intensive computations in the Elemental library.
This intensive computetions can be done distributed in multiple nodes.

Data from R can be sent to Elemental and then retrieved after the computation

To run in multiple nodes MPI is required, so every node runs R and Elemental
}

