% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{RowSwap}
\alias{RowSwap}
\title{Swaps two rows in an Elemental matrix}
\usage{
RowSwap(MatrixA, to, from)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}

\item{to}{row destination (input) 0-indexed}

\item{to}{row origin (input) 0-indexed}
}
\value{
None
}
\description{
Interchanges two rows inside an Elemental matrix
}

