% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Scale}
\alias{Scale}
\title{Multiplies an Elemental matrix by an scalar}
\usage{
Scale(alpha, MatrixA)
}
\arguments{
\item{alpha}{a scalar value}

\item{MatrixA}{an Elemental matrix (input)}
}
\value{
None
}
\description{
Multiplies each element of the MatrixA by an scalar alpha
}

