% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ScaleTrapezoid}
\alias{ScaleTrapezoid}
\title{Multiplies a trapeoidal section of an Elemental matrix by an scalar}
\usage{
ScaleTrapezoid(alpha, uplo, MatrixA, offset)
}
\arguments{
\item{alpha}{a scalar value (input)}

\item{uplo}{(input). It can take the values "U" or "L" to indicate what part of
the matrix should be modified.}

\item{MatrixA}{an Elemental matrix (input, output)}

\item{offset}{an integer value that indicates the diagonal offset (input)}
}
\value{
None
}
\description{
Multiplies each element of the MatrixA by an scalar alpha
}

