% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Shift}
\alias{Shift}
\title{Shifts an Elemental matrix by an scalar}
\usage{
Shift(MatrixA, alpha)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input)}

\item{alpha}{a scalar value (input)}
}
\value{
None
}
\description{
Adds a scalar value to each element of the MatrixA
}

