% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ShiftDiagonal}
\alias{ShiftDiagonal}
\title{Shifts the diagonal elements of an Elemental matrix by an scalar}
\usage{
ShiftDiagonal(MatrixA, alpha, offset = 0)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}

\item{alpha}{a scalar value (input)}

\item{offset}{an integer value that indicates the diagonal offset (input)}
}
\value{
None
}
\description{
Adds a scalar value to the diafonal elements of the MatrixA
}

