% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Sin}
\alias{Sin}
\title{Element-wise sine on a Elemental matrix}
\usage{
Sin(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Replaces each entry from MatrixA by its sine function
}

