% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Sqrt}
\alias{Sqrt}
\title{Element-wise square root on a Elemental matrix}
\usage{
Sqrt(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Replaces each entry from MatrixA by its square root
}

