% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Swap}
\alias{Swap}
\title{Swaps two Elemental matrices}
\usage{
Swap(orientation, MatrixX, MatrixY)
}
\arguments{
\item{orientation}{indicates whether the matrix is transposed or no (input).
Possible values "N" or "T" Normal or Transposed}

\item{MatrixX}{an Elemental matrix (input, output)}

\item{MatrixY}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Interchanges the contents of two elemental matrices, MatrixX becomes
MatrixY and vice-versa
}

