% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SymmetricMax}
\alias{SymmetricMax}
\title{Finds the maximum value in an Elemental symmetric matrix}
\usage{
SymmetricMax(uplo, MatrixA)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" (upper or lower)
to indicate in what part of the matrix is the data.}

\item{MatrixA}{an Elemental matrix (input)}
}
\value{
$value the maximum value in the matrix

  	   $indices an array with the coordinates of the maximum value
}
\description{
Finds both, the maximum scalar value in a symmetric matrix and its coordinates (row, col)
}

