% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SymmetricMaxAbs}
\alias{SymmetricMaxAbs}
\title{Finds the maximum absolute value in an Elemental symmetric matrix}
\usage{
SymmetricMaxAbs(uplo, MatrixA)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" (upper or lower)
to indicate in what part of the matrix is the data.}

\item{MatrixA}{an Elemental matrix (input)}
}
\value{
$value the maximum absolute value in the matrix

  	   $indices an array with the coordinates of the maximum value
}
\description{
Finds both, the maximum absolute scalar value in a 
symmetric matrix and its coordinates (row, col)
}

