% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SymmetricMin}
\alias{SymmetricMin}
\title{Finds the minimum value in an Elemental symmetric matrix}
\usage{
SymmetricMin(uplo, MatrixA)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" (upper or lower)
to indicate in what part of the matrix is the data.}

\item{MatrixA}{an Elemental matrix (input)}
}
\value{
$value the minimum value in the matrix

  	   $indices an array with the coordinates of the minimum value
}
\description{
Finds both, the minimum scalar value in a symmetric matrix and its coordinates (row, col)
}

