% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SymmetricMinAbs}
\alias{SymmetricMinAbs}
\title{Finds the minimum absolute value in an Elemental symmetric matrix}
\usage{
SymmetricMinAbs(uplo, MatrixA)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" (upper or lower)
to indicate in what part of the matrix is the data.}

\item{MatrixA}{an Elemental matrix (input)}
}
\value{
$value the minimum absolute value in the matrix

  	   $indices an array with the coordinates of the minimum value
}
\description{
Finds both, the minimum absolute scalar value in a 
symmetric matrix and its coordinates (row, col)
}

