% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Symv}
\alias{Symv}
\title{Symmetric matrix vector multiplication of Elemental matrices}
\usage{
Symv(uplo, alpha, MatrixA, Matrixx, beta, Matrixy)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" to indicate what part of
matrixA has entries.}

\item{alpha}{Scalar that multiplies the matrix vector product(input)}

\item{MatrixA}{a hermitian Elemental matrix (input)}

\item{Matrixx}{a 1-D Elemental matrix (input)}

\item{beta}{Scalar that multiplies scales Matrixy (input)}

\item{Matrixy}{a 1-D Elemental matrix (input, output)}
}
\value{
None
}
\description{
Performs the operation `y = aAx + by` on Elemental matrices
}

