% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Syr2}
\alias{Syr2}
\title{Symmetric rank two update on Elemental matrices}
\usage{
Syr2(uplo, alpha, Matrixx, Matrixy, MatrixA)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" to indicate what part of
matrixA has entries.}

\item{alpha}{Scalar that multiplies the outer products (input)}

\item{Matrixx}{a 1-D Elemental matrix (input)}

\item{Matrixy}{a 1-D Elemental matrix (input)}

\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Performs the operation `A = alpha (xy^T + yx^T) + A`
on Elemental matrices, where only the selected section of A is updated
}

