% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Tan}
\alias{Tan}
\title{Element-wise tangent on a Elemental matrix}
\usage{
Tan(MatrixA)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Replaces each entry from MatrixA by its tangent function
}

