% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Transpose}
\alias{Transpose}
\title{Transposes an Elemental matrix}
\usage{
Transpose(MatrixA, MatrixB)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input)}

\item{MatrixB}{an Elemental matrix (input, output)}
}
\value{
None
}
\description{
Transposes the MatrixA and stores into MatrixB
}

