% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Trsv}
\alias{Trsv}
\title{Triangular solve for Elemental matrices}
\usage{
Trsv(uplo, orientation, diagUnit, MatrixA, Matrixx)
}
\arguments{
\item{uplo}{(input). It can take the values "U" or "L" to indicate what part of
matrixA has entries.}

\item{orientation}{indicates whether matrixA is transposed or no (input).
Possible values "N" or "T" Normal or Transposed}

\item{diagUnit}{indicates if matrixA has a unit diagonal; true or false (input)}

\item{MatrixA}{an Elemental matrix (input)}

\item{Matrixx}{a 1-D Elemental matrix (input, output)}
}
\value{
None
}
\description{
Performs the operation `x = A^(-1)x` on Elemental matrices. A is treated
as either lower or upper triangular.
}

