% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Uniform}
\alias{Uniform}
\title{Fills an Elemental matrix with an uniform distribution}
\usage{
Uniform(MatrixA, rows, cols, center = 0, radius = 1)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input and output)}

\item{rows}{the number of rows (input)}

\item{cols}{the number of columns (input)}

\item{center}{the center of the uniform distribution (input)}

\item{radius}{the radius of the uniform distribution (input)}
}
\value{
None
}
\description{
Fills MatrixX with uniformly distributed entries
}

