% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{WorldSize}
\alias{WorldSize}
\title{MPI Wold Size}
\usage{
WorldSize()
}
\value{
an integer with the number of processors
}
\description{
Returns the number of MPI processes involved in the computation
}

