% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Zeros}
\alias{Zeros}
\title{Fills an Elemental matrix with zeoros}
\usage{
Zeros(MatrixA, rows, cols)
}
\arguments{
\item{MatrixA}{an Elemental matrix (input and out)}

\item{rows}{the number of rows (input, output)}

\item{cols}{the number of columns (input, output)}
}
\value{
None
}
\description{
Fills with zeros and modifies the dimensions of the Elemental matrix
MatrixX
}

