% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{Dot}
\alias{Dot}
\title{Dot Product for Elemental Matrices}
\usage{
Dot(MatrixA, MatrixB)
}
\arguments{
\item{MatrixA}{An Elemental Matrix (input)}

\item{MatrixB}{An Elemental Matrix (input)}
}
\value{
A scalar value
}
\description{
Perform the dot product beween MatrixA and Matrix B returning an scalar
depending on the matrix datatype
}

