% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ElDistMatrix-class}
\alias{ElDistMatrix}
\alias{ElDistMatrix-class}
\title{Class to represent an elemental Dist Matrix}
\description{
Class to represent an elemental Dist Matrix
}
\section{Slots}{

\describe{
\item{\code{ptr}}{A pointer to the distributed matrix object (C++)}

\item{\code{datatype}}{A character that stores the datatype}

\item{\code{active}}{a boolean value that indicates whether the grid is
still alive or not.}
}}

