% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Environment.R
\name{Initialize}
\alias{Initialize}
\title{Initialize}
\usage{
Initialize()
}
\value{
None
}
\description{
Initializes MPI allowing the use of Distributed Matrices
}
\examples{
Initialize()
A<-DistMatrix()
Uniform(A,2,2)
Print(A)
Finalize()
}

